# backend/internal/docs/swagger.yaml
basePath: /api/
definitions:
  _ResponseWithData:
    description: Общий ответ success/error, содержащий произвольные данные.
    properties:
      data:
        description: Объект полезной нагрузки
      status:
        description: Результат запроса
        type: string
    type: object
  _ResponseWithMessage:
    description: Общий простой ответ, который передает только понятное для человека
      сообщение.
    properties:
      message:
        description: Человеко-читаемое сообщение
        type: string
      status:
        description: Результат запроса
        type: string
    type: object
...
  hackathon-back_internal_model.FAQ:
    properties:
      answer:
        example: Для восстановления пароля используйте форму 'Забыли пароль' на странице
          входа.
        type: string
      category:
        example: authentication
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: 1a2b3c4d-5678-90ab-cdef-1234567890ab
        type: string
      id:
        example: 7b2aab2e-4d1f-45b5-90c5-4d5d4db5ef11
        type: string
      is_active:
        example: true
        type: boolean
      order:
        example: 1
        type: integer
      question:
        example: Как восстановить пароль?
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  hackathon-back_internal_model.FAQCategoryResponse:
    properties:
      category:
        type: string
      faqs:
        items:
          $ref: '#/definitions/hackathon-back_internal_model.FAQ'
        type: array
    type: object
  hackathon-back_internal_model.FAQCreateRequest:
    properties:
      answer:
        example: Для восстановления пароля используйте форму 'Забыли пароль' на странице
          входа.
        type: string
      category:
        example: authentication
        type: string
      order:
        example: 1
        type: integer
      question:
        example: Как восстановить пароль?
        type: string
    required:
    - answer
    - category
    - question
    type: object
  hackathon-back_internal_model.FAQListResponse:
    properties:
      faqs:
        items:
          $ref: '#/definitions/hackathon-back_internal_model.FAQ'
        type: array
      total:
        type: integer
    type: object
  hackathon-back_internal_model.FAQUpdateRequest:
    properties:
      answer:
        example: Используйте кнопку 'Забыли пароль' на странице входа.
        type: string
      category:
        example: auth
        type: string
      is_active:
        example: false
        type: boolean
      order:
        example: 2
        type: integer
      question:
        example: Как сбросить пароль?
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Флоу авторизации: сначала пользователь регистрируется, в ответе получает модельку своего user и токен.
    Токен + код с почты он отправляет на ручку confirm, если что-то идёт не так (токен истёк, не пришёл код на почту), то запрос нужно отправить на ручку resend-confirmation.
    Далее уже можно авторизоваться, login/refresh/test-login выставляет в cookie access и refresh токены, фронтенду, ничего с ними делать не нужно, они сами по себе живут в браузере и отправляются при каждом запросе.
    Специально для мобильного приложения при login/refresh/test-login токены дублируются в теле ответа.
    При запросе к защищённым ручкам API мобильному приложению необходимо выставить заголовок Authorization: Bearer *access_token*.
    При refresh мобильное приложение передаёт refresh токен в теле запроса.
  title: Hackathon API
  version: 0.1.0
paths:
...
  /faq:
    get:
      description: Возвращает список часто задаваемых вопросов с пагинацией и фильтрацией
      parameters:
      - description: Фильтр по категории
        in: query
        name: category
        type: string
      - description: Фильтр по активности
        in: query
        name: is_active
        type: boolean
      - default: 50
        description: Лимит (по умолчанию 50, максимум 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список FAQ
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/hackathon-back_internal_model.FAQListResponse'
              type: object
        "400":
          description: Некорректные параметры запроса
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при получении списка FAQ
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить список FAQ
      tags:
      - FAQ
    post:
      consumes:
      - application/json
      description: Создает новый часто задаваемый вопрос
      parameters:
      - description: Данные для создания FAQ
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hackathon-back_internal_model.FAQCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: FAQ успешно создан
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/hackathon-back_internal_model.FAQ'
              type: object
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при создании FAQ
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Создать новый FAQ
      tags:
      - FAQ
  /faq/{id}:
    delete:
      description: Удаляет часто задаваемый вопрос
      parameters:
      - description: FAQ UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: FAQ успешно удалён
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Неверный параметр пути
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: FAQ не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при удалении FAQ
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Удалить FAQ
      tags:
      - FAQ
    get:
      description: Возвращает часто задаваемый вопрос по его ID
      parameters:
      - description: FAQ UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные FAQ
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/hackathon-back_internal_model.FAQ'
              type: object
        "400":
          description: Неверный параметр пути
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: FAQ не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при получении FAQ
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить FAQ по ID
      tags:
      - FAQ
    patch:
      consumes:
      - application/json
      description: Обновляет данные часто задаваемого вопроса
      parameters:
      - description: FAQ UUID
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hackathon-back_internal_model.FAQUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: FAQ успешно обновлён
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: FAQ не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при обновлении FAQ
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Обновить FAQ
      tags:
      - FAQ
  /faq/categories:
    get:
      description: Возвращает список всех категорий FAQ
      produces:
      - application/json
      responses:
        "200":
          description: Список категорий
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Ошибка при получении категорий
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить список категорий
      tags:
      - FAQ
  /faq/category/{category}:
    get:
      description: Возвращает активные часто задаваемые вопросы по указанной категории
      parameters:
      - description: Категория FAQ
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список FAQ категории
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/hackathon-back_internal_model.FAQ'
                  type: array
              type: object
        "400":
          description: Неверный параметр пути
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при получении FAQ
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить FAQ по категории
      tags:
      - FAQ
  /faq/grouped:
    get:
      description: Возвращает все активные FAQ сгруппированные по категориям
      produces:
      - application/json
      responses:
        "200":
          description: FAQ по категориям
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/hackathon-back_internal_model.FAQCategoryResponse'
                  type: array
              type: object
        "500":
          description: Ошибка при получении данных
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить FAQ по категориям
      tags:
      - FAQ
...
swagger: "2.0"